some restructuring in app
