+++ app/src/main/java/com/example/gainly/MainActivity.kt
@@ -4,8 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.example.gainly.com.example.gainly.App
-import com.example.gainly.ui.theme.GainlyTheme
+import com.example.gainly.presentation.ui.theme.GainlyTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
