+++ app/src/main/java/com/example/gainly/common/Resource.kt
@@ -1,0 +1,7 @@
+package com.example.gainly.common
+
+sealed class Resource<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T) : Resource<T>(data)
+    class Error<T>(message: String, data: T? = null) : Resource<T>(data, message)
+    class Loading<T>(data: T? = null) : Resource<T>(data)
+}
