+++ app/src/main/java/com/example/gainly/presentation/community/CommunityScreen.kt
@@ -1,0 +1,13 @@
+package com.example.gainly.presentation.community
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun CommunityScreen(
+    modifier: Modifier = Modifier
+) {
+    Text("CommunityScreen")
+}
+
