+++ app/src/main/java/com/example/gainly/presentation/home/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.example.gainly
+package com.example.gainly.presentation.home
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -8,8 +8,5 @@ import androidx.compose.ui.Modifier
 fun HomeScreen(
     modifier: Modifier = Modifier
 ) {
-    Text(
-        text = "HomeScreen",
-        modifier = modifier
-    )
+    Text("HomeScreen")
 }
