+++ app/src/main/java/com/example/gainly/presentation/navigation/models/TopLevelRoute.kt
@@ -1,0 +1,10 @@
+package com.example.gainly.presentation.navigation.models
+
+import androidx.compose.ui.graphics.vector.ImageVector
+import com.example.gainly.presentation.navigation.TopLevelRoutes
+
+data class TopLevelRoute(
+    val name: String,
+    val route: TopLevelRoutes,
+    val icon: ImageVector
+)
